Demo_App/requirements.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
gradio==4.44.1
pillow==10.4.0
pandas==2.0.3
numpy==1.24.3
opencv-python==4.9.0.80
matplotlib==3.7.2
scikit-learn==1.3.0 